Fix multi-env raycaster viewer plugin race condition.
