+++ genesis/vis/viewer_plugins/__init__.py
@@ -1,7 +1,9 @@
-from .plugins import *
-from .viewer_plugin import (
+from .base import (
     EVENT_HANDLE_STATE,
     EVENT_HANDLED,
-    RaycasterViewerPlugin,
     ViewerPlugin,
 )
+from .raycast import (
+    RaycasterViewerPlugin,
+)
+from .plugins import *
